Allow ApplicationSystemTestCase driver override
Allow for the local developer to provide an override of the default
ApplicationSystemTestCase driver of :headless_chrome via the
BROWSER_TEST_DRIVER environment variable. This allows for the one-off
ability to interact with the browser in non-headless mode for
troubleshooting and debugging while keeping the default behavior
unchanged.

This will use the default :headless_chrome driver:
rails test:system

This will override the default and use :chrome for the current test run:
BROWSER_TEST_DRIVER=chrome rails test:system

Completes #10
Allow for ApplicationSystemTest driver to be provided by CLI