Fix Devise views for Tailwind and Turbo
The generated Devise views looked terrible with the default
implementation of Tailwind in Rails. Made the following improvements to
make them look better with Tailwind:

* Took all of the views and nested them within div tags
* Added basic button and link styling
* Added basic header styling

In order to not repeat the exact same strings over and over again,
created a TailwindHelper class that can be used to produce desired class
combinations from helper methods.

Fixed the button_to in registrations/edit so that the confirm works with
Turbo.

Completes Fix Devise views to work with Tailwind subtask on #14 Users
need to be able to sign up and authenticate