Add system tests for log in and log out
Add tests for log in and log out scenarios as well as login errors.
Because there was some repeated code in the sign up page, refactored out
the log in logic into helper methods in ApplicationSystemTestCase that
allow for login as a user or login with specific credentials. Also had
to add a constant to TestHelper so that I could access User passwords
in the ApplicationSystemTestCase login as user method.

Completes User can login and logout subtask on #14 Users need to be able
to sign up and authenticate

Contributes to:
Error handling (invalid password, signup error) works as intended
subtask