Allow time for emails to be sent in system tests
In certain situations in the Devise tests, there were tests failing
frequently or intermittently because not enough time elapsed following a
line of code where an email was sent before the next line of code was
executed. This seems to be isolated to system tests, so created a
method in ApplicationSystemTestCase that yields the block and then
sleeps.

Modified the Devise test cases where emails are being sent so they
enclose actions that trigger emails in an await_jobs block.

Also specify patch version on GitHub action so that it won't fail when a
new patch version of Ruby is released.

Fixes #18 Tests that depend upon emails being sent can fail if not
enough time elapses before assertion that depends upon it