Refactor TailwindHelper to stop at border
Refactor the implementation of the TailwindHelper methods to stop at the
border of the element and leave the implementation of margin and
positioning to the particular view. This prevents the issue of trying to
take away margin you don't want from an included component.

Also did some overall refactoring of the module by splitting it into
submodules and adding a tw_base_button method that the other button
methods inherit from.

Modified the views to use the modified methods and have them handle the
margin.

Completes #24 Refactor Tailwind helper code