Set Devise to paranoid mode and adjust tests
Set Devise to paranoid mode to prevent against malicious users receiving
feedback about whether an account exists in the database or not.

See Testing for User Enumeration and Guessable User Account
(OWASP-AT-002) and Devise paranoid mode on Wiki for more information.

Uncommented paranoid setting to turn on paranoid mode and adjusted
existing Devise system tests to account for the new behavior. Put an
item in the backlog to re-evaluate whether to change the locking and
unlocking messaging to account for paranoid mode.

Completes #25 Make Devise paranoid