Add contributing guidelines to the repo
Add CONTRIBUTING and BURN_THE_CONTRIBUTOR_COVENANT_WITH_FIRE markdown
files to the repository with guidelines for collaboration and
contribution to the project. Add internal links within the document
using GitHub markdown.

This will likely be tweaked in the next commit when we update our
README.

Completes #27 Add a CONTRIBUTING.md file to the repo