Update bundle and fix new rubocop offenses
Update the bundle to clear Dependabot alerts on the app, including an
upgrade to Rails 7.0.4.2. Since the bundle update contained new RuboCop
cops and we have them enabled by default, corrected the newly appearing
violations. Corrections include:

* Adding a blank line before assert statements in tests
* Removal of redundant :: in constant notation
* Removal of redundant escapes in regular expressions
* Added missing documentation comment for ApplicationMailer

The bundle update caused the GitHub action build to fail with
error: Downloading debug-1.7.1 revealed dependencies not in the API or
the lockfile.
  (irb (>= 1.5.0), reline (>= 0.3.1)).

These dependencies are not writing to the Gemfile.lock when the install
happens. Resolved by placing a pessimistic version constraint on debug
to keep it at 1.6.

Completes #33 Update bundle and Rails to >= 7.0.4.1 to clear Dependabot
alerts