Upgrade Devise to 4.9 and decommission hacks
Upgrade Devise (4.9.0) and Responders (3.1.0) gems to their current
versions and follow the upgrade instructions on the Devise wiki to
upgrade the application to 4.9.

There was one change to a failure message for blank email that created a
new test failure that needed to be corrected. The new behavior is
satisfactory so we just updated the assertions.

Summary of Changes:
* Update Devise in Gemfile and `bundle update devise` to upgrade the
  Responders gem
* Add the configuration for responder error_status to
  :unprocessable_entity and redirect_status to :see_other in the Devise
  initializer file
* Remove the temporary TurboFailureApp from the initializer
* Remove the parent_controller config from the initializer
* Remove the failure app warden config from initializer
* Delete the TurboDeviseController that is no longer necessary
* Update LogInLogOutTest to use new error message for blank email and
  refactor into instance variable

Completes #38 Decommision Devise Turbo hacks once an official solution
is in place