Upgrade Ruby to 3.4.3 and get tests passing
Upgrade the Ruby version from 3.2.2 to 3.4.3 in the app and upgrade
Rails to the latest patch version of 7.0. In upgrading from Ruby 3.2 to
Ruby 3.4, there were some libraries that were removed from the standard
library to become separate gems. These gems were temporarily added to
the Gemfile so that the app could boot and tests would pass with
Rails 7.0 and Ruby 3.4. They will hopefully be removed in subsequent
commits as we upgrade Rails beyone 7.0.

Temporary Gemfile adds:
- `bigdecimal`
- `drb`
- `mutex_m`

Rubocop target Ruby version was held at 3.2 because the other gems have
not yet been updated.

Also replaced `actions/checkout@v2` with `actions/checkout@v4` in the
GitHub action.

Completes sub-task Upgrade Ruby to latest version for issue Upgrade to
Rails 8 and current versions of Gems #51