Upgrade Tailwind from version 3 to 4
Upgrade the app to move from TailwindCSS version 3 to version 4. Before
making the change, upgraded tailwindcss-rails to latest version of 3
(3.3.2). Since this app doesn't have a package.json, had to temporarily
add one. The upgrade script now works locally without needing to install
the npm packages globally. The next step should have been to upgrade the
tailwindcss-rails version in the bundle to 4.x and run the
`bin/rails tailwindcss:upgrade` command, but instead ran the
`npx @tailwindcss/upgrade` command.

Note: Even though the app didn't have a node_modules directory
associated with this commit, it was necessary to add it to the
.gitignore file or the upgrade script fails.

After completing the upgrade, deleted the package.json,
package-lock.json, node_modules/ and restored .gitignore to its previous
state.

After running the upgrade, was running into issues with the
stylesheet_link_tag for Tailwind in the application.css file and getting
the framework to load correctly. Re-ran the
`bin/rails tailwindcss:install` command to have the application.html.erb
and bin/dev files modified.

After testing, temporarily added back in the compatibility styles for
borders and re-tested the application.

Completes sub-task Upgrade Tailwind from 3 to 4 for issue Upgrade to
Rails 8 and current versions of Gems #51