Upgrade app from Rails 7.0 to 7.1
Upgrade the bundle for Rails from the latest 7.0 version to the latest
7.1 version and run the upgrade script. Allowed all of the overwrites,
with the exception of adding back in the email hosts in the development
and test environment files.

Instead of enabling the new environment defaults one by one, modified
the configuration file to use the new defaults for 7.1 and tested the
app with automation and manual regression.

See the following resources:
- Upgrading Ruby on Rails guide (7.0 to 7.1 section in particular)
  https://guides.rubyonrails.org/upgrading_ruby_on_rails.html
- Rails 7.1 release notes
  https://guides.rubyonrails.org/7_1_release_notes.html

Completes sub-task Upgrade Rails from 7.0 to 7.1 for issue Upgrade to
Rails 8 and current versions of Gems #51