Upgrade app from Rails 7.1 to 7.2
Upgrade the bundle for Rails from the latest 7.1 version to the latest
7.2 version and run the upgrade script. Allowed all of the overwrites,
with the exception of adding back in the email hosts in the development
and test environment files. Also made the following changes:

- Removed the temporary gems from the Gemfile that were added in to get
  Rails 7.0 working with Ruby 3.4
- Added the allow_browser versions: :modern constraint to
  ApplicationController

Instead of enabling the new environment defaults one by one, modified
the configuration file to use the new defaults for 7.2 and tested the
app with automation and manual regression.

See the following resources:
- Upgrading Ruby on Rails guide (7.1 to 7.2 section in particular)
  https://guides.rubyonrails.org/upgrading_ruby_on_rails.html
- Rails 7.2 release notes
  https://guides.rubyonrails.org/7_2_release_notes.html

Completes sub-task Upgrade Rails from 7.1 to 7.2 for issue Upgrade to
Rails 8 and current versions of Gems #51