Upgrade app from Rails 7.2 to 8.0
Upgrade the bundle for Rails from the latest 7.2 version to the latest
8.0 version and run the upgrade script. Allowed all of the overwrites,
with the exception of adding back in the email hosts in the development
and test environment files.

Instead of enabling the new environment defaults one by one, modified
the configuration file to use the new defaults for 8.0 and tested the
app with automation and manual regression.

See the following resources:
- Upgrading Ruby on Rails guide (7.2 to 8.0 section in particular)
  https://guides.rubyonrails.org/upgrading_ruby_on_rails.html
- Rails 8.0 release notes
  https://guides.rubyonrails.org/8_0_release_notes.html

Completes sub-task Upgrade Rails from 7.2 to 8.0 for issue Upgrade to
Rails 8 and current versions of Gems #51

Also found no breaking changes in Devise or had any oustanding framework
defaults to enable and will move the following sub-tasks to done:
- Ensure Devise still works (will be a subset of each upgrade and a more
  thorough review at the end) to done.
- Resolve any outstanding upgrade default configurations